+++ modules/api-export/src/main/java/com/axway/apim/api/export/lib/APIDeleteCLIOptions.java
@@ -13,8 +13,9 @@ public class APIDeleteCLIOptions extends APIExportCLIOptions {
 
 	@Override
 	public void printUsage(String message, String[] args) {
-		super.printUsage(message, args);		
-		System.out.println("You may run one of the following examples:");
+		super.printUsage(message, args);
+		System.out.println("----------------------------------------------------------------------------------------");
+		System.out.println("How to delete APIs using different filter options:");
 		System.out.println(getBinaryName()+" api delete -s api-env");
 		System.out.println(getBinaryName()+" api delete -s api-env -n \"*API*\"");
 		System.out.println(getBinaryName()+" api delete -s api-env -id f6106454-1651-430e-8a2f-e3514afad8ee");