+++ modules/api-import/src/test/java/com/axway/apim/test/basic/NoAPIDefinitionConfiguredIT.java
@@ -33,7 +33,7 @@ public class NoAPIDefinitionConfiguredIT extends TestNGCitrusTestRunner {
 		echo("####### Calling the tool with a Non-Admin-User. #######");
 		createVariable(ImportTestAction.API_DEFINITION,  "");
 		createVariable(ImportTestAction.API_CONFIG,  "/com/axway/apim/test/files/basic/minimal-config.json");
-		createVariable("status", "unpublished");
+		createVariable("state", "unpublished");
 		createVariable("expectedReturnCode", String.valueOf(ErrorCode.NO_API_DEFINITION_CONFIGURED.getCode()));
 		swaggerImport.doExecute(context);
 	}