+++ modules/apim-adapter/src/main/java/com/axway/apim/api/APIBaseDefinition.java
@@ -6,8 +6,6 @@ package com.axway.apim.api;
  * @author cwiechmann@axway.com
  */
 public class APIBaseDefinition extends API {
-	
-	private String createdOn;
 
 	@Override
 	public String getState() {
@@ -16,13 +14,5 @@ public class APIBaseDefinition extends API {
 
 	public void setStatus(String state) {
 		this.state = state;
-	}
-	
-	public String getCreatedOn() {
-		return createdOn;
-	}
-
-	public void setCreatedOn(String createdOn) {
-		this.createdOn = createdOn;
 	}
 }