+++ modules/apim-adapter/src/main/java/com/axway/apim/lib/APIMCoreCLIOptions.java
@@ -8,12 +8,9 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Option;
-import org.apache.commons.cli.OptionGroup;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 
-import com.axway.apim.lib.CommandParameters;
-import com.axway.apim.lib.RelaxedParser;
 import com.axway.apim.lib.errorHandling.ErrorCode;
 
 public abstract class APIMCoreCLIOptions {