+++ modules/apim-adapter/src/main/java/com/axway/apim/lib/StandardExportCLIOptions.java
@@ -15,12 +15,18 @@ public abstract class StandardExportCLIOptions extends APIMCoreCLIOptions {
 		option.setRequired(false);
 		options.addOption(option);
 		
-		option = new Option("deleteFolder", "Controls if an existing local folder should be deleted. Defaults to false.");
+		option = new Option("deleteTarget", "Controls if an existing target folder or file should be deleted. Defaults to false.");
 		options.addOption(option);
 		
-		option = new Option("f", "format", true, "Controls the export format. By default the console is used.");
+		option = new Option("t", "target", true, "Defines the target location for get operations that are creating files or directories.\n"
+				+ "Defaults to current folder. Required output files or directories are created automatically.");
 		option.setRequired(false);
-		option.setArgName("console|json");
+		option.setArgName("my/apis");
+		options.addOption(option);
+		
+		option = new Option("o", "output", true, "Controls the output format. By default the console is used.");
+		option.setRequired(false);
+		option.setArgName("console|json|csv");
 		options.addOption(option);
 	}
 }